requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
executing==0.8.3
filelock==3.6.0
fonttools==4.33.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.5.1
hyperframe==5.2.0
idna==2.10
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
preshed==3.0.6
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
rfc3986==1.5.0
sacremoses==0.0.49
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
thinc==8.0.15
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
